chore: organize protocol frontend models
